Split out factories to not fall foul of function limit
